demos/python_demos/stable_diffusion/requirements.txt
git+https://github.com/huggingface/optimum-intel.git
onnx
tritonclient[grpc]
openvino==2023.2
pillow
